Squashed commit of the following:
commit 2a5326ee04aaded51f7f36f42f2a409a04df5928
Merge: 6f882530 d62b40f8
Author: Tom Clarke <tomcl@users.noreply.github.com>
Date:   Fri Aug 19 16:32:08 2022 +0100

    Merge branch 'master' into tc-dev

commit 6f882530a4c32dab53350e70ba4b6bac2b0068f1
Author: Tom Clarke <tomcl@users.noreply.github.com>
Date:   Fri Aug 19 16:20:32 2022 +0100

    mend divider bar, improve waveform column width calculation

commit 326c85ac6244cd808b4f180c02d7ba6bfa7addff
Author: Tom Clarke <tomcl@users.noreply.github.com>
Date:   Fri Aug 19 13:56:59 2022 +0100

    finish work on wavesim RAMs

commit 0309c3fd53b781664c3a32045d4cab604614335a
Author: Tom Clarke <tomcl@users.noreply.github.com>
Date:   Thu Aug 18 23:01:57 2022 +0100

    checkpoint RAM changes in waveform simulator

+++ src/Renderer/Common/CommonTypes.fs
@@ -635,7 +635,7 @@ module CommonTypes
         /// Radix in which values are displayed in the wave simulator
         Radix: NumberBase option
         /// Width of the waveform column
-        WaveformColumnWidth: int option
+        WaveformColumnWidth: float option
         /// Number of visible cycles in the waveform column
         ShownCycles: int option
         /// RAMs which are selected to be shown in the RAM tables
